+++ CHANGES.md
@@ -5,7 +5,9 @@
 ### Bug Fixes
 
 * filter: Improved warning and error messages in the case of missing columns. [#1604] (@victorlin)
+* merge: Non-id columns in metadata inputs that would conflict with the output id column are now forbidden and will cause an error if present.  Previously they would overwrite values in the output id column, causing incorrect output. [#1593][] (@tsibley)
 
+[#1593]: https://github.com/nextstrain/augur/pull/1593
 [#1604]: https://github.com/nextstrain/augur/pull/1604
 
 ## 25.3.0 (22 August 2024)